requirements.txt
asgiref==3.2.3
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
get==2019.4.13
gunicorn==20.0.4
post==2019.4.13
public==2019.4.13
PyJWT==1.7.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
psycopg2==2.8.4